Rename verification module
