+++ art/classifiers/lightgbm.py
@@ -106,7 +106,7 @@ class LightGBMClassifier(Classifier, ClassifierDecisionTree):
         :return: A list of decision trees.
         :rtype: `[Tree]`
         """
-        from art.metrics.metrics_trees import Box, Tree
+        from art.metrics.verification_decisions_trees import Box, Tree
 
         booster_dump = self._model.dump_model()["tree_info"]
         trees = list()
@@ -126,7 +126,7 @@ class LightGBMClassifier(Classifier, ClassifierDecisionTree):
 
     def _get_leaf_nodes(self, node, i_tree, class_label, box):
         from copy import deepcopy
-        from art.metrics.metrics_trees import LeafNode, Box, Interval
+        from art.metrics.verification_decisions_trees import LeafNode, Box, Interval
 
         leaf_nodes = list()
